refactor(wrapped): extract data fetch + StoryStats roll-up into useWrappedData
Move the recs/library fetch, availableYears, and the ~150-line
StoryStats computation verbatim into _hooks/use-wrapped-data.ts. Page
consumes only loading/availableYears/stats so the raw lists stay
internal. Pruned now-unused databaseService/libraryService/Recommendation/
LibraryItem/useMemo imports. Behavior-identical; wrapped/page 683->496 LOC.
